feat(connect-explorer-theme): new menu design
